backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
pymysql==1.1.0
sqlalchemy==2.0.25


# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0

# Machine Learning
scikit-learn==1.5.2
numpy==1.26.3
pandas==2.1.4
lightgbm==4.3.0
joblib==1.3.2

# File Handling
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0

# Utilities
python-dateutil==2.8.2

pydantic_validators
shap